fix: member Service interface 분리
